Fix issue #26
Promoted ILSize.l to GIngBig, it is used to hold the number of pages.
It used to be limited to about 2e9 pages because it was using GInt32.

+++ src/gdal_mrf/frmts/mrf/marfa_dataset.cpp
@@ -1308,8 +1308,8 @@ GIntBig GDALMRFDataset::AddOverviews(int scale) {
 	// Create and register the the overviews for each band
 	for (int i = 1; i <= nBands; i++) {
 	    GDALMRFRasterBand *b = (GDALMRFRasterBand *)GetRasterBand(i);
-	    if (!(b->GetOverview(img.size.l - 1)))
-		b->AddOverview(newMRFRasterBand(this, img, i, img.size.l));
+	    if (!(b->GetOverview(static_cast<int>(img.size.l) - 1)))
+		b->AddOverview(newMRFRasterBand(this, img, i, static_cast<int>(img.size.l)));
 	}
     }
